requirements.txt
langchain
langchain-google-genai
gspread
python-dotenv
langsmith
streamlit
pandas
oauth2client